refactor: fold TaskLoader into task registry functions
